requirements.txt
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
gunicorn==21.2.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.31.0
httpx>=0.24.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Pillow==10.4.0